[fix and improvement]
before

In current document, It said "이 함수 표현은 메소드 함수가 아닌 곳에 가장 적합합니다. 그래서 생성자로서 사용할 수 없습니다."
It means
" This function expression is best suited for non-method functions. That's why you can't use it as a constructor."
But it is not.
As I know Arrow function does not have a prototype that's why it can't be used as a constructor.
this translation needs to be fixed.
And some of the information was not translated into Korean

after

* Invalid information about Arrowfunction was removed.
* Added omitted information.